Fix a bunch of different issues, most of which relating to the SPS: - Fix the SPS sometimes not being able to insert energy due to the order in which things tick #6351 The SPS is designed to be able to use all energy inserted, even if some of it may end up being unused - Fix the SPS marking the amount of input processed by more than it actually was able to process (due to the fact that there was not that much input stored) - Fix the SPS displaying the max theoretical process rate (if the input tank had unlimited capacity), instead of displaying the actual amount it processed - Fix SPS Process Rate having wrong capitalization for mB - Fix a few mismatched equals/hashcode implementations - Fix MathUtils clampToLong converting it to an int not a long

+++ src/api/java/mekanism/api/math/MathUtils.java
@@ -49,7 +49,7 @@ public class MathUtils {
      */
     public static long clampToLong(double d) {
         if (d < Long.MAX_VALUE) {
-            return (int) d;
+            return (long) d;
         }
         return Long.MAX_VALUE;
     }
